Add a base CreateFromUploadView class
